Ajoute la consigne audio de la situation livraison.
