+++ src/app/situation_livraison.js
@@ -2,11 +2,11 @@ import 'commun/infra/report_erreurs';
 
 import { afficheSituation } from 'commun/vues/affiche_situation';
 
-import DepotRessourcesQuestions from 'questions/infra/depot_ressources_questions';
+import DepotRessourcesLivraison from 'livraison/infra/depot_ressources_livraison';
 import Situation from 'commun/modeles/situation';
 import AdaptateurVueSituation from 'questions/vues/situation';
 
 const situation = new Situation({ modeEntrainement: true });
 
-const depotRessources = new DepotRessourcesQuestions();
+const depotRessources = new DepotRessourcesLivraison();
 afficheSituation('livraison', situation, AdaptateurVueSituation, depotRessources);
